VERY WIP - beginning refactor of scheduler for pluggability
