+++ scheduler/backends/kubernetes/pods_test.go
@@ -1,4 +1,4 @@
-package scheduler
+package kubernetes
 
 import (
 	"testing"
