+++ scheduler/scheduler.go
@@ -6,8 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"path"
-	"reflect"
 	"strconv"
 	"strings"
 	"sync"
@@ -19,61 +17,48 @@ import (
 	logrus "github.com/sirupsen/logrus"
 	"golang.org/x/crypto/ssh"
 
+	"k8s.io/client-go/kubernetes"
+
 	nats "github.com/nats-io/nats.go"
 	config "github.com/nre-learning/antidote-core/config"
 	"github.com/nre-learning/antidote-core/db"
 	models "github.com/nre-learning/antidote-core/db/models"
 	"github.com/nre-learning/antidote-core/services"
-
 	// Custom Network CRD Types
-	networkcrd "github.com/nre-learning/antidote-core/pkg/apis/k8s.cni.cncf.io/v1"
-
 	// Kubernetes Types
-	corev1 "k8s.io/api/core/v1"
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	rest "k8s.io/client-go/rest"
-
 	// Kubernetes clients
-
-	kubernetesCrd "github.com/nre-learning/antidote-core/pkg/client/clientset/versioned"
-	kubernetesExt "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	kubernetes "k8s.io/client-go/kubernetes"
 )
 
-const initContainerName string = "copy-local-files"
+// TODO - figure out if this is the right place to insert this abstraction, or if we should do it lower.
+// Also the most sticky thing you're likely to experience is the ease of use offered by k8s in deleting a namespace. See if Openstack has a similar property, or punt
+// this to the implementor of that plugin
+// Need to go through the full deletion and GC process and ensure it's not k8s specific anywhere
+//
+// This interface will obviously be a requirement, but you'll have opinions on what people do within each fulfillment function. For instance, updating test state as endpoints come online.
+// This opens up an interesting question about whether or not those should be defined in an interface. Bottom line, you'll have plugins requirements that go beyond
+// implementing these high-level functions, so you should think about those  and document them. What kind of DB functions need to be called, and when?
+type SchedulerBackend interface {
 
-// NetworkCrdClient is an interface for the client for our custom
-// network CRD. Allows for injection of mocks at test time.
-type NetworkCrdClient interface {
-	UpdateNamespace(string)
-	Create(obj *networkcrd.NetworkAttachmentDefinition) (*networkcrd.NetworkAttachmentDefinition, error)
-	Update(obj *networkcrd.NetworkAttachmentDefinition) (*networkcrd.NetworkAttachmentDefinition, error)
-	Delete(name string, options *meta_v1.DeleteOptions) error
-	Get(name string) (*networkcrd.NetworkAttachmentDefinition, error)
-	List(opts meta_v1.ListOptions) (*networkcrd.NetworkList, error)
+	// TODO - this is likely not meant to be implemented in the interface. Each plugin will need an Init function
+	// that returns a ready-to-go instance. This will allow for publicly set attributes (antidoteocnfig) but also allow internal types to be set (kubeconfig, etc)
+	Init() error
+
+	handleRequestCREATE(ot.SpanContext, services.LessonScheduleRequest)
+	handleRequestMODIFY(ot.SpanContext, services.LessonScheduleRequest)
+	handleRequestBOOP(ot.SpanContext, services.LessonScheduleRequest) // TODO - consider if this is necessary anymore. The API has the state, no need to update k8s or other backend labels
+	handleRequestDELETE(ot.SpanContext, services.LessonScheduleRequest)
+
+	configureStuff(ot.SpanContext, models.LiveLesson, services.LessonScheduleRequest)
+	PurgeOldLessons(ot.SpanContext)
+	PurgeOldSessions(ot.SpanContext)
 }
 
 // AntidoteScheduler is an Antidote service that receives commands for things like creating new lesson instances,
 // moving existing livelessons to a different stage, deleting old lessons, etc.
 type AntidoteScheduler struct {
-	KubeConfig    *rest.Config
-	NC            *nats.Conn
-	Config        config.AntidoteConfig
-	Db            db.DataManager
-	HealthChecker LessonHealthChecker
-
-	// Allows us to disable GC for testing. Production code should leave this at
-	// false
-	// DisableGC bool
-
-	// Client for interacting with normal Kubernetes resources
-	Client kubernetes.Interface
-
-	// Client for creating CRD defintions
-	ClientExt kubernetesExt.Interface
-
-	// Client for creating instances of our network CRD
-	ClientCrd kubernetesCrd.Interface
+	NC     *nats.Conn
+	Config config.AntidoteConfig
+	Db     db.DataManager
 
 	BuildInfo map[string]string
 }
@@ -82,11 +67,13 @@ type AntidoteScheduler struct {
 // and put a results message on the "results" channel when finished (success or fail)
 func (s *AntidoteScheduler) Start() error {
 
-	// Ensure our network CRD is in place (should fail silently if already exists)
-	err := s.createNetworkCrd()
-	if err != nil {
-		return err
-	}
+	// TODO - you moved much of the scheduler state to the plugin, so you'll need to copy the logic that the cmd/antidoted/ main is doing, into here based on plugin
+
+	// In case we're restarting from a previous instance, we want to make sure we clean up any
+	// orphaned k8s namespaces by killing any with our ID. This should be done synchronously
+	// before the scheduler or the API is started.
+	logrus.Info("Pruning orphaned namespaces...")
+	scheduler.PruneOrphanedNamespaces()
 
 	// Garbage collection
 	go func() {
@@ -94,7 +81,7 @@ func (s *AntidoteScheduler) Start() error {
 
 			span := ot.StartSpan("scheduler_gc")
 			// Clean up any old lesson namespaces
-			llToDelete, err := s.PurgeOldLessons(span.Context())
+			llToDelete, err := s.PurgeOldLessons(span.Context()) // TODO - this is in namespaces.go
 			if err != nil {
 				span.LogFields(log.Error(err))
 				ext.Error.Set(span, true)
@@ -155,55 +142,7 @@ func (s *AntidoteScheduler) Start() error {
 	return nil
 }
 
-// PurgeOldSessions cleans up old LiveSessions according to the configured LiveSessionTTL
-func (s *AntidoteScheduler) PurgeOldSessions(sc ot.SpanContext) error {
-	span := ot.StartSpan("scheduler_purgeoldsessions", ot.ChildOf(sc))
-	defer span.Finish()
-
-	lsList, err := s.Db.ListLiveSessions(span.Context())
-	if err != nil {
-		span.LogFields(log.Error(err))
-		ext.Error.Set(span, true)
-		return err
-	}
-
-	lsTTL := time.Duration(s.Config.LiveSessionTTL) * time.Minute
-
-	for _, ls := range lsList {
-		createdTime := time.Since(ls.CreatedTime)
-
-		// No need to continue if this session hasn't even exceeded the TTL
-		if createdTime <= lsTTL {
-			continue
-		}
-
-		llforls, err := s.getLiveLessonsForSession(span.Context(), ls.ID)
-		if err != nil {
-			span.LogFields(log.Error(err))
-			ext.Error.Set(span, true)
-			return err
-		}
-
-		// We don't want/need to clean up this session if there are active livelessons that are using it.
-		if len(llforls) > 0 {
-			continue
-		}
-
-		// TODO(mierdin): It would be pretty rare, but in the event that a livelesson is spun up between the request above
-		// and the livesession deletion below, we would encounter the leak bug we saw in 0.6.0. It might be worth seeing if
-		// you can lock things somehow between the two.
-
-		err = s.Db.DeleteLiveSession(span.Context(), ls.ID)
-		if err != nil {
-			span.LogFields(log.Error(err))
-			ext.Error.Set(span, true)
-			return err
-		}
-	}
-
-	return nil
-}
-
+// TODO - this should be in the DB package probably?
 func (s *AntidoteScheduler) getLiveLessonsForSession(sc ot.SpanContext, lsID string) ([]string, error) {
 	span := ot.StartSpan("scheduler_getlivelessonsforsession", ot.ChildOf(sc))
 	defer span.Finish()
@@ -232,161 +171,8 @@ func (s *AntidoteScheduler) getLiveLessonsForSession(sc ot.SpanContext, lsID str
 	return retLLIDs, nil
 }
 
-func (s *AntidoteScheduler) configureStuff(sc ot.SpanContext, nsName string, ll models.LiveLesson, newRequest services.LessonScheduleRequest) error {
-	span := ot.StartSpan("scheduler_configure_stuff", ot.ChildOf(sc))
-	defer span.Finish()
-	span.SetTag("llID", ll.ID)
-	span.LogFields(log.Object("llEndpoints", ll.LiveEndpoints))
-
-	s.killAllJobs(span.Context(), nsName, "config")
-
-	wg := new(sync.WaitGroup)
-	wg.Add(len(ll.LiveEndpoints))
-	allGood := true
-	for i := range ll.LiveEndpoints {
-
-		// Ignore any endpoints that don't have a configuration option
-		if ll.LiveEndpoints[i].ConfigurationType == "" {
-			wg.Done()
-			continue
-		}
-
-		job, err := s.configureEndpoint(span.Context(), ll.LiveEndpoints[i], newRequest)
-		if err != nil {
-			span.LogFields(log.Error(err))
-			ext.Error.Set(span, true)
-			return err
-		}
-
-		go func() {
-			defer wg.Done()
-
-			oldStatusCount := map[string]int32{
-				"active":    0,
-				"succeeded": 0,
-				"failed":    0,
-			}
-			for i := 0; i < 600; i++ {
-				completed, statusCount, err := s.getJobStatus(span, job, newRequest)
-				if err != nil {
-					allGood = false
-					return
-				}
-
-				// The use of this map[string]int32 and comparing old with new using DeepEqual
-				// allows us to only log changes in status, rather than the periodic spam
-				if !reflect.DeepEqual(oldStatusCount, statusCount) {
-					span.LogFields(
-						log.String("jobName", job.Name),
-						log.Object("statusCount", statusCount),
-					)
-				}
-
-				if completed {
-					return
-				}
-				oldStatusCount = statusCount
-				time.Sleep(1 * time.Second)
-			}
-			allGood = false
-			return
-		}()
-
-	}
-
-	wg.Wait()
-
-	if !allGood {
-		return errors.New("Problem during configuration")
-	}
-
-	return nil
-}
-
-// getVolumesConfiguration returns a slice of Volumes, VolumeMounts, and init containers that should be used in all pod and job definitions.
-// This allows Syringe to pull lesson data from either Git, or from a local filesystem - the latter of which being very useful for lesson
-// development.
-func (s *AntidoteScheduler) getVolumesConfiguration(sc ot.SpanContext, lessonSlug string) ([]corev1.Volume, []corev1.VolumeMount, []corev1.Container, error) {
-	span := ot.StartSpan("scheduler_get_volumes", ot.ChildOf(sc))
-	defer span.Finish()
-
-	lesson, err := s.Db.GetLesson(span.Context(), lessonSlug)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	volumes := []corev1.Volume{}
-	volumeMounts := []corev1.VolumeMount{}
-	initContainers := []corev1.Container{}
-
-	// Init container will mount the host directory as read-only, and copy entire contents into an emptyDir volume
-	initContainers = append(initContainers, corev1.Container{
-		Name:  initContainerName,
-		Image: "bash",
-		Command: []string{
-			"bash",
-		},
-
-		// In previous versions of this platform, we used the subPath parameter of the VolumeMount to specify the actual lesson directory (i.e. lessons/my-new-lesson) to make available
-		// to the lesson endpoints at the /antidote location. This would result in something like /antidote/<lesson files> rather than /antidote/lessons/my-new-lesson/<lesson files>,
-		// which was much more convenient to access within the lessons.
-		//
-		// However, some runtimes don't appear to handle subPath well, as shown here: https://github.com/kata-containers/runtime/issues/2812.
-		//
-		// Fortunately, we don't **actually** need the subPath field to accomplish the same goal. This field seems to be mostly useful in environments where you want
-		// to use the same volume but provide different mount points to different containers or pods. In our case, all pods within a lesson should look the same, and
-		// volumes are created at a pod level using emptyDir. So, instead of using subPath in the mount, we can just copy the correct subdirectory here in the init container,
-		// and mount the whole volume in the main container. This volume will already be prepped with the relevant subdirectory.
-		Args: []string{
-			"-c",
-			fmt.Sprintf(
-				"ls -lha /antidote-ro && cp -r /antidote-ro/%s/* /antidote && adduser -D antidote && chown -R antidote:antidote /antidote && ls -lha /antidote",
-				strings.TrimPrefix(lesson.LessonDir, fmt.Sprintf("%s/", path.Clean(s.Config.CurriculumDir))),
-			),
-		},
-
-		VolumeMounts: []corev1.VolumeMount{
-			{
-				Name:      "host-volume",
-				ReadOnly:  true,
-				MountPath: "/antidote-ro",
-			},
-			{
-				Name:      "local-copy",
-				ReadOnly:  false,
-				MountPath: "/antidote",
-			},
-		},
-	})
-
-	// Add outer host volume, should be mounted read-only
-	volumes = append(volumes, corev1.Volume{
-		Name: "host-volume",
-		VolumeSource: corev1.VolumeSource{
-			HostPath: &corev1.HostPathVolumeSource{
-				Path: s.Config.CurriculumDir,
-			},
-		},
-	})
-
-	// Add inner container volume, should be mounted read-write so we can copy files into it
-	volumes = append(volumes, corev1.Volume{
-		Name: "local-copy",
-		VolumeSource: corev1.VolumeSource{
-			EmptyDir: &corev1.EmptyDirVolumeSource{},
-		},
-	})
-
-	// Finally, mount local copy volume as read-write
-	volumeMounts = append(volumeMounts, corev1.VolumeMount{
-		Name:      "local-copy",
-		ReadOnly:  false,
-		MountPath: "/antidote",
-	})
-
-	return volumes, volumeMounts, initContainers, nil
-
-}
+// TODO - all of the reachable stuff below should be mandated somehow. The question is, should this be moved into the backend but mandated through interfaces,
+// or centralized, and mandate that all of the plugins to use it via documented convention?
 
 func (s *AntidoteScheduler) isEpReachable(ep *models.LiveEndpoint) (bool, error) {
